namespacing/nesting, using docblock comments, refactoring
