+++ Build1/zmachine/zmachine/ConsoleIO.cs
@@ -1,7 +1,6 @@
-﻿using System;
-
-namespace zmachine
+﻿namespace zmachine
 {
+    using System;
     public class ConsoleIO : IIO
     {
         public string ReadLine()
