+++ Build1/zmachine/zmachine/Machine.cs
@@ -1,33 +1,47 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-
-namespace zmachine
+﻿namespace zmachine
 {
+    using System;
+    using System.Collections.Generic;
+    using System.Diagnostics;
 
+    /// <summary>
+    /// This class moves through the input file and extracts bytes to deconstruct instructions in the code
+    /// </summary>
     public partial class Machine
     {
         public const int StackDepth = 128;
         public const int MemorySize = 1024 * 128;
+        /// <summary>
+        /// Stack of size 32768 (can be larger, but this should be fine)
+        /// </summary>
         public const int StackSize = 1024 * 32;
 
-        // This class moves through the input file and extracts bytes to deconstruct instructions in the code
-        private readonly Memory memory = new Memory(size: MemorySize);         // Initialize memory
-        private readonly Memory stack = new Memory(size: StackSize);           // Stack of size 32768 (can be larger, but this should be fine)
+        private readonly Memory memory = new Memory(size: MemorySize);
+        private readonly Memory stack = new Memory(size: StackSize);
         private readonly ObjectTable objectTable;
         private readonly IIO io;
         private readonly Lex lex;
 
         private readonly bool debug = false;
 
-        //  StateOfPlay stateofplay = new StateOfPlay();    // Will contain (1) Local variable table, (2) contents of the stack, (3) value of PC, (4) current routine call state.
-        private uint programCounter = 0;                                        // Program Counter to step through memory
-        private uint pcStart = 0;                                   // Program Counter at the beginning of executing the instruction
-        private bool finish = false;                                // Flag to say "finish processing. We're done".
+        /// <summary>
+        /// Program Counter to step through memory
+        /// </summary>
+        private uint programCounter = 0;
+        /// <summary>
+        /// Program Counter at the beginning of executing the instruction
+        /// </summary>
+        private uint pcStart = 0;
+        /// <summary>
+        /// Flag to say "finish processing. We're done".
+        /// </summary>
+        private bool finish = false;
         private uint stackPointer = 0;
         private uint callDepth = 0;
-        RoutineCallState[] callStack = new RoutineCallState[StackDepth];  // We could use a "Stack" here, as well, but we have lots of memory. According to the spec, there could be up to 90. But 128 is nice.
-
+        /// <summary>
+        /// We could use a "Stack" here, as well, but we have lots of memory. According to the spec, there could be up to 90. But 128 is nice.
+        /// </summary>
+        private readonly RoutineCallState[] callStack = new RoutineCallState[StackDepth];
         // Class constructor : Loads in data from file and sets Program Counter
         public Machine(IIO io, string programFilename)
         {
@@ -121,7 +135,9 @@ namespace zmachine
             setVar(pc_getByte(), returnVal);                                // Set the return value. Calling a routine is a "store" function, so the next byte contains where to store the result.
         }
 
-        // Find the PC start point in the header file and set PC 
+        /// <summary>
+        /// Find the PC start point in the header file and set PC 
+        /// </summary>
         public void setProgramCounter()
         {
             programCounter = memory.getWord(Memory.ADDR_INITIALPC);               // Set PC by looking at pc start byte in header file
@@ -133,8 +149,9 @@ namespace zmachine
         //      - 0OP
         //      - VAR
 
-
-        // Looks at pointer and returns instruction
+        /// <summary>
+        /// Looks at pointer and returns instruction
+        /// </summary>
         public void processInstruction()
         {
             pcStart = programCounter;
@@ -301,103 +318,6 @@ namespace zmachine
             op.run(this, operands);
         }
 
-        private readonly OpcodeHandler_2OP[] OP2opcodes = {
-        new  op_unknown_2op(),      // OPCODE/HEX
-        new  op_je(),               // 01/01 a b ?(label)
-        new  op_jl(),               // 02/02 a b ?(label)	
-        new  op_jg(),               // 03/03
-        new  op_dec_chk(),          // 04/04
-        new  op_inc_chk(),          // 05/05
-        new  op_jin(),              // 06/06
-        new  op_test(),             // 07/07
-        new  op_or(),               // 08/08
-        new  op_and(),              // 09/09
-        new  op_test_attr(),        // 10/0A
-        new  op_set_attr(),         // 11/0B
-        new  op_clear_attr(),       // 12/0C
-        new  op_store(),            // 13/0D
-        new  op_insert_obj(),       // 14/0E
-        new  op_loadw(),            // 15/0F
-        new  op_loadb(),            // 16/10
-        new  op_get_prop(),         // 17/11
-        new  op_get_prop_addr(),    // 18/12
-        new  op_get_next_addr(),    // 19/13
-        new  op_add(),              // 20/14
-        new  op_sub(),              // 21/15
-        new  op_mul(),              // 22/15
-        new  op_div(),              // 23/16
-        new  op_mod(),              // 24/17
-        new  op_unknown_2op(),      // 25/19 
-        new  op_unknown_2op(),      // 26/1A
-        new  op_unknown_2op(),      // 27/1B 
-        new  op_unknown_2op(),      // 28/1C 
-        new  op_unknown_2op(),      // 29/1D 
-        new  op_unknown_2op(),      // 30/1E
-        new  op_unknown_2op()       // 31/1F 
-     };
-        private readonly OpcodeHandler_1OP[] OP1opcodes = {
-                                    // OPCODE/HEX
-        new  op_jz(),               // 128/01
-        new  op_get_sibling(),      // 129/01 	
-        new  op_get_child(),        // 130/02 
-        new  op_get_parent(),       // 131/03
-        new  op_get_prop_len(),     // 132/04
-        new  op_inc(),              // 133/05
-        new  op_dec(),              // 134/06
-        new  op_print_addr(),       // 135/07
-        new  op_unknown_1op(),      // 136/08
-        new  op_remove_obj(),       // 137/09
-        new  op_print_obj(),        // 138/0A
-        new  op_ret(),              // 139/0B
-        new  op_jump(),             // 140/0C
-        new  op_print_paddr(),      // 141/0D
-        new  op_load(),             // 142/0E
-        new  op_unknown_1op(),      // 141/0F
-        new  op_unknown_1op()       // 142/0F
-        };
-        private readonly OpcodeHandler_0OP[] OP0opcodes = {
-                                    // OPCODE/HEX
-        new  op_rtrue(),            // 176/00 
-        new  op_rfalse(),           // 177/01 
-        new  op_print (),           // 178/02 
-        new  op_print_ret(),        // 179/03
-        new  op_nop(),              // 180/04
-        new  op_save (),            // 181/05
-        new  op_restore (),         // 182/06
-        new  op_restart(),          // 183/07
-        new  op_ret_popped(),       // 184/08
-        new  op_pop(),              // 185/09
-        new  op_quit(),             // 186/0A
-        new  op_new_line(),         // 187/0B
-        new  op_show_status(),      // 188/0C
-        new  op_verify ()           // 189/0C
-        };
-        private readonly OpcodeHandler_OPVAR[] VARopcodes = {
-                                    // OPCODE/HEX
-        new  op_call(),             // 224/00 
-        new  op_storew(),           // 225/01 
-        new  op_storeb(),           // 226/02
-        new  op_put_prop(),         // 227/03
-        new  op_sread(),            // 228/04
-        new  op_print_char(),       // 229/05
-        new  op_print_num(),        // 230/06
-        new  op_random(),           // 231/07
-        new  op_push(),             // 232/08
-        new  op_pull(),             // 233/09
-        new  op_split_window(),     // 234/0A
-        new  op_set_window(),       // 235/0B
-        new  op_unknown_op_var(),   // 236/1E
-        new  op_unknown_op_var(),   // 237/1E
-        new  op_unknown_op_var(),   // 238/1E
-        new  op_unknown_op_var(),   // 239/1E
-        new  op_unknown_op_var(),   // 240/1E
-        new  op_unknown_op_var(),   // 241/1E
-        new  op_unknown_op_var(),   // 242/1E
-        new  op_output_stream(),    // 243/13
-        new  op_input_stream()      // 244/14
-        };
-        // --------------------------------------------------------------------------------------------
-
         public void branch(bool condition)        // Check branch instructions
         {
             //             Debug.WriteLine("BRANCH: " + (condition ? "1" : "0"));
@@ -482,49 +402,5 @@ namespace zmachine
             return value;
 
         }
-
-        public CPUState State
-        {
-            get
-            {
-                return new CPUState(
-                    memory: this.memory.Contents,
-                    stack: this.stack.Contents,
-                    lexMemoryPointer: this.lex.MemoryPointer,
-                    pc: this.programCounter,
-                    pcStart: this.pcStart,
-                    sp: this.stackPointer,
-                    callDepth: this.callDepth,
-                    callStack: this.callStack,
-                    finish: this.finish);
-            }
-            set
-            {
-                this.memory.load(value.memory);
-                this.stack.load(value.stack);
-                this.lex.MemoryPointer = value.lexMemoryPointer;
-                this.programCounter = value.programCounter;
-                this.pcStart = value.pcStart;
-                this.stackPointer = value.stackPointer;
-                this.callDepth = value.callDepth;
-                this.callStack = value.callStack;
-                this.finish = value.finish;
-            }
-        }
-        // --------------------------------------------------------------------------------------------
-
-
-
-
-        public string stateString()
-        {
-            string s = "M: " + memory.getCrc32() + " S: " + stack.getCrc32();
-            //            for (ushort i = 1; i < 256; ++i)
-            //                s += " " + getVar(i);
-            return s;
-        }
-        //        --------------------------------------------------------------------------------------------
-
-
     } // end Machine
 } // end namespace
